chore: Prevent tortoise from running in multiple event loops
