+++ src/pwncore/__init__.py
@@ -1,5 +1,7 @@
 from fastapi import FastAPI
 
+from tortoise.contrib.fastapi import register_tortoise
+
 import pwncore.docs as docs
 import pwncore.routes as routes
 
@@ -8,3 +10,11 @@ app = FastAPI(
 )
 
 app.include_router(routes.router)
+
+register_tortoise(
+    app,
+    db_url="sqlite://:memory:",
+    modules={"models": ["pwncore.db"]},
+    generate_schemas=True,
+    add_exception_handlers=True
+)
